backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
databases[sqlite]==0.8.0
sqlalchemy==1.4.53
pydantic[email]==1.10.13
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai==0.3.2
supabase==2.3.0
# Google Calendar API
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-auth==2.22.0